Move tests for LoadingSpinner and Navbar components
