+++ frontend/src/__tests__/components/LoadingSpinner.test.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react'
 import { render, screen } from '@testing-library/react'
 import { ThemeProvider, createTheme } from '@mui/material/styles'
-import LoadingSpinner from '../components/LoadingSpinner'
+import LoadingSpinner from '../../components/LoadingSpinner'
 
 const renderWithTheme = (mode: 'light' | 'dark', ui: ReactNode = <LoadingSpinner />) => {
   const theme = createTheme({
